requirements.txt
numpy
pandas
pillow
tqdm
pydicom
scikit-learn
scipy
scikit-image
opencv-python
torch
torchvision
torchaudio
transformers
qwen_vl_utils
timm
matplotlib
seaborn
open-clip-torch
